backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.4.26
colorama==0.4.6
distro==1.9.0
Django==5.2
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
lxml==5.4.0
openai==1.82.1
packaging==25.0
pillow==11.2.1
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.9.0
PyPDF2==3.0.1
python-docx==1.1.2
python-pptx==1.0.2
setuptools==80.9.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
wheel==0.45.1
whitenoise==6.9.0
XlsxWriter==3.2.3
